Naprawiona baza

+++ src/main/java/hello/AlbumRepo.java
@@ -2,5 +2,5 @@ package hello;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface AlbumRepo extends CrudRepository <Album,Long> {
+public interface AlbumRepo extends CrudRepository <Album,Integer> {
 }
